+++ CHANGES.md
@@ -4,6 +4,7 @@
 
 - **BREAKING CHANGE**: Renaming of `rasters(_rio).get_nodata_value` function to `rasters(_rio).get_nodata_value_from_dtype`. Older function is deprecated.
 - **ENH: Creating `rasters.get_nodata_value_from_xr` to retrieve the nodata value as specified by `rioxarray`**
+- FIX: Fix handling of nodata with xarray structures in `sertit.rasters`
 - FIX: Handle None cases in `ci.assert_val`
 
 ## 1.40.0 (2024-08-05)